More changes for clean compilation

+++ DeepSkyStacker/RegisterSettings.cpp
@@ -270,7 +270,7 @@ void RegisterSettings::on_percentStack_textEdited(const QString &text)
 
 void RegisterSettings::on_luminanceThreshold_valueChanged(int newValue)
 {
-	if (detectionThreshold != newValue)
+	if (detectionThreshold != static_cast<uint>(newValue))
 	{
 		detectionThreshold = newValue;
 		// Display new value
